% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_sample.R
\name{bootstrap_sample}
\alias{bootstrap_sample}
\title{Generate boostrap sample on the provided population.}
\usage{
bootstrap_sample(pop, n, n_bootstraps)
}
\arguments{
\item{pop}{Population data}

\item{n}{Sample sizes (SBS sample size, PPS sample size).}

\item{n_bootstraps}{Number of bootstrap samples.}
}
\value{
A summary data frame of the estimator.
}
\description{
Generate boostrap sample on the provided population.
}
