% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_agreement_weights.R
\name{calculate_agreement_weights}
\alias{calculate_agreement_weights}
\title{Calculate Agreement Weights}
\usage{
calculate_agreement_weights(ranks, set_size)
}
\arguments{
\item{set_size}{Set size}

\item{RV}{Ranking values}
}
\value{
A vector of agreement weights
}
\description{
Calculate Agreement Weights
}
\keyword{internal}
