% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoefF.R
\name{calculate_coefficients}
\alias{calculate_coefficients}
\title{This function computes the coefficient of variance estimator}
\usage{
calculate_coefficients(H, n)
}
\arguments{
\item{H}{Set size for each ranking group.}

\item{n}{Sample size.}
}
\description{
This function computes the coefficient of variance estimator
}
\keyword{internal}
