% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inclusion_probability.R
\name{calculate_first_order_prob}
\alias{calculate_first_order_prob}
\title{Calculate first order inclusion probability.}
\usage{
calculate_first_order_prob(i, size_measurement, n, total_size)
}
\arguments{
\item{i}{Index of the probability being calculated.}

\item{size_measurement}{Size measurements of population units.}

\item{n}{Sample sizes (SBS sample size, PPS sample size).}

\item{total_size}{Sum of size measurement.}
}
\value{
A vector of inclusion probabilities.
}
\description{
Calculate first order inclusion probability.
}
\keyword{internal}
