% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inclusion_probability.R
\name{calculate_inclusion_prob}
\alias{calculate_inclusion_prob}
\title{Calculate inclusion probabilities.}
\usage{
calculate_inclusion_prob(size_measurement, n, parallelize = TRUE)
}
\arguments{
\item{size_measurement}{Size measurements of population units.}

\item{n}{Sample sizes (SBS sample size, PPS sample size).}

\item{parallelize}{A flag whether to parallelize the computational task.}
}
\value{
A vector of inclusion probabilities.
}
\description{
Calculate inclusion probabilities.
}
