% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoefF.R
\name{CoefF}
\alias{CoefF}
\title{This function computes the coefficient of variance estimator}
\usage{
CoefF(H, n)
}
\arguments{
\item{H}{The set size}

\item{n}{The sample size}
}
\value{

}
\description{
This function computes the coefficient of variance estimator
}
\keyword{internal}
