% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emergence_ranks}
\alias{emergence_ranks}
\title{Ranks for Seed Emergence.}
\format{
A data frame with 6 variables: \code{seed_emergence}, \code{ranker1}, \code{ranker2},
\code{ranker3}, \code{ranker4} and \code{ranker5}.
}
\usage{
emergence_ranks
}
\description{
Contains the ranks given by 5 rankers of the number of seeds emerged in a sample of 15 plots.
}
\keyword{datasets}
