% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_empirical_population.R
\name{get_empirical_population}
\alias{get_empirical_population}
\title{Compute empirical population by imputing reponses and measured sizes using knn.}
\usage{
get_empirical_population(sample_indices, pop, y)
}
\arguments{
\item{sample_indices}{Indices of sample.}

\item{pop}{Population data frame to be sampled with 4 columns.
\enumerate{
\item Halton numbers
\item X1-coordinate of population unit
\item X2-coordinate of population unit
\item Size measurement of population unit
}}

\item{y}{Sample response values.}
}
\value{
A summary data frame of the estimator.
}
\description{
Compute empirical population by imputing reponses and measured sizes using knn.
}
\keyword{internal}
