% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbs_pps_sample.R
\name{get_sbs_pps_sample_indices}
\alias{get_sbs_pps_sample_indices}
\title{Generate SBS PPS sample indices.}
\usage{
get_sbs_pps_sample_indices(population, n, with_unique_pps = FALSE)
}
\arguments{
\item{population}{Population data frame to be sampled with 2 columns.
\enumerate{
\item Halton numbers
\item Size measurements of population units
}}

\item{n}{Sample sizes (SBS sample size, PPS sample size).}

\item{with_unique_pps}{A boolean to specify whether to use unique indices for pps sample or not.}
}
\value{
A named list of:
\itemize{
\item sbs_pps_indices: sample indices
\item sbs_indices: sbs sample indices
\item pps_indices: pps sample indices
\item sizes_wo_sbs: measured sizes without sbs sample
}
}
\description{
Generate SBS PPS sample indices.
}
\keyword{internal}
