% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_empirical_population.R
\name{impute_w_knn}
\alias{impute_w_knn}
\title{Calculate the mean of the nearest neighbors.}
\usage{
impute_w_knn(knn_row, y)
}
\arguments{
\item{knn_row}{A row of knn matrix composed of indices and distances.}

\item{y}{Sample response values.}
}
\value{
A summary data frame of the estimator.
}
\description{
Calculate the mean of the nearest neighbors.
}
\keyword{internal}
