% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InPlotSampling-package.R
\docType{package}
\name{InPlotSampling}
\alias{InPlotSampling-package}
\alias{InPlotSampling}
\title{InPlotSampling: Easing the Application of Ranked Set Sampling in Practice}
\description{
The InPlotSampling package provides a way for researchers to easily implement Ranked Set Sampling in practice. Ranked Set Sampling was originally described by McIntyre (1952) (reprinted in 2005) \doi{10.1198/000313005X54180}. This package takes work by Omer and Kravchuk (2021) \url{https://doi.org/10.1007/s13253-021-00439-1} and enables easy use of the methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://biometryhub.github.io/InPlotSampling/}
  \item Report bugs at \url{https://github.com/biometryhub/InPlotSampling/issues}
}

}
\author{
\strong{Maintainer}: Sam Rogers \email{sam.rogers@adelaide.edu.au}

Authors:
\itemize{
  \item Omer Ozturk \email{omer@stat.osu.edu} (\href{https://orcid.org/0000-0002-0235-4772}{ORCID})
  \item Olena Kravchuk \email{olena.kravchuk@adelaide.edu.au} (\href{https://orcid.org/0000-0001-5291-3600}{ORCID}) [data contributor]
  \item Peter Kasprzak \email{peter.kasprzak@adelaide.edu.au}
}

}
