% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jps_estimator.R
\name{jps_estimator}
\alias{jps_estimator}
\title{Computes the estimator for JPS data}
\usage{
jps_estimator(data, set_size, replace = TRUE, model_based, N, alpha)
}
\arguments{
\item{data}{The data to use for estimation.}

\item{set_size}{Set size for each raking group.}

\item{replace}{Logical (default \code{TRUE}). Sample with replacement?}

\item{model_based}{An inference mode:
\itemize{
\item \code{FALSE}: design based inference
\item \code{TRUE}: model based inference using super population model
}}

\item{N}{The population size.}

\item{alpha}{The significance level.}
}
\value{
A \code{data.frame} with the point estimates provided by JPS estimators along with standard error and
confidence intervals.
}
\description{
Computes the estimator for JPS data
}
