% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jps_estimator_single.R
\name{jps_estimator_single}
\alias{jps_estimator_single}
\title{Computes the estimator and variance for each individual ranker}
\usage{
jps_estimator_single(
  ranks,
  y,
  set_size,
  N,
  coef,
  coef_del,
  replace,
  model_based,
  K
)
}
\arguments{
\item{ranks}{Ranks of Y.}

\item{y}{Response measurements.}

\item{set_size}{Set size for each raking group.}

\item{N}{Finite population size.}

\item{coef}{Coefficients used in variance computation when sample size is n.}

\item{coef_del}{Coefficients used in variance computation when the i-th unit is deleted.}

\item{replace}{Logical. Sample with replacement?}

\item{model_based}{An inference mode:
\itemize{
\item \code{FALSE}: design based inference
\item \code{TRUE}: model based inference using super population model
}}

\item{K}{Number of rankers.}
}
\value{
A \code{data.frame} with the point estimates provided by JPS estimators along with standard error and
confidence intervals.
}
\description{
Computes the estimator and variance for each individual ranker
}
\keyword{internal}
