% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JPSD0F.R
\name{JPSD0F}
\alias{JPSD0F}
\title{Title}
\usage{
JPSD0F(pop, sample_size, set_size, number_rankers, tau)
}
\arguments{
\item{pop}{A vector of values which is the population samples are taken from.}

\item{sample_size}{The size of the sample to be taken.}

\item{number_rankers}{The number of rankers.}

\item{tau}{The ranking quality.}
}
\value{

}
\description{
Title
}
