% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JPSD2F.R
\name{JPSD2F}
\alias{JPSD2F}
\title{Generate JPS sampling without replacement on the provided population.}
\usage{
JPSD2F(pop, n, H, tau, K, with_replacement = FALSE)
}
\arguments{
\item{tau}{A parameter which controls ranking quality.}
}
\value{
A matrix with ranks from each ranker.
}
\description{
Generate JPS sampling without replacement on the provided population.
}
