% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JPSED0F.R
\name{JPSED0F}
\alias{JPSED0F}
\title{Title}
\usage{
JPSED0F(RV, Y, set_size, Coef, N, Replace, Model)
}
\arguments{
\item{RV}{Ranking values provided}

\item{Y}{The data values that were ranked}

\item{set_size}{The set}

\item{Model}{}
}
\value{
A vector with two elements: the mean and variance estimate from JPS
}
\description{
Title
}
\keyword{internal}
