% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JPSEF.R
\name{JPSEF}
\alias{JPSEF}
\title{Computes the estimator for JPS data}
\usage{
JPSEF(data, setsize, replace = TRUE, model, N, alpha)
}
\arguments{
\item{data}{The data to use for estimation}

\item{setsize}{The set size}

\item{replace}{Logical (default \code{TRUE}). Sample with replacement?}

\item{N}{The population size.}

\item{alpha}{The significance level.}
}
\value{

}
\description{
Computes the estimator for JPS data
}
\keyword{internal}
