% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JPSLF.R
\name{JPSLF}
\alias{JPSLF}
\title{Title}
\usage{
JPSLF(
  data,
  set_size,
  replace,
  population_size,
  model,
  estimator = "SD Weighted"
)
}
\arguments{
\item{data}{Data to use to estimate.}

\item{set_size}{Number of samples which are ranked in each set.}

\item{replace}{Logical. Are samples drawn with replacement?}

\item{population_size}{Population size of the population which \code{data} is taken from.}

\item{model}{}
}
\value{

}
\description{
Title
}
