% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListF.R
\name{ListF}
\alias{ListF}
\title{Title}
\usage{
ListF(data, set_size, Replace, N, Model, CoefD0)
}
\arguments{
\item{data}{The data}

\item{set_size}{set size}

\item{Replace}{replacement}

\item{N}{Population size}

\item{Model}{Model}

\item{CoefD0}{Coef}
}
\value{

}
\description{
Title
}
\keyword{internal}
