% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{Seed Emergence Population.}
\format{
A data frame with 2640 rows of 2 variables: \code{actual_seed_emergence} and
\code{estimated_seed_emergence} giving the actual and estimated number of seeds emerged.
}
\usage{
population
}
\description{
The entire population of actual and estimated seed emergence.
}
\keyword{datasets}
