% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_estimate_single.R
\name{rss_estimate_single}
\alias{rss_estimate_single}
\title{This function Computes RSS estimator and its variance without replacement sampling}
\usage{
rss_estimate_single(first_ranker_ranks, y, set_size, pop_size)
}
\arguments{
\item{first_ranker_ranks}{Ranks from the first ranker.}

\item{y}{Response variable.}

\item{set_size}{Set size.}

\item{pop_size}{Population size.}
}
\description{
This function Computes RSS estimator and its variance without replacement sampling
}
\keyword{internal}
