% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_jps_estimate.R
\name{rss_jps_estimate}
\alias{rss_jps_estimate}
\title{Estimate means from RSS or JPS sample.}
\usage{
rss_jps_estimate(
  data,
  set_size,
  method,
  confidence = 0.95,
  replace = TRUE,
  model_based = FALSE,
  pop_size = NULL
)
}
\arguments{
\item{data}{A data frame of \code{JPS} or \code{RSS} rankings.}

\item{set_size}{The set size of the ranks.}

\item{method}{A method used to sample:
\itemize{
\item \code{"JPS"}: Judgment-post stratified sampling
\item \code{"RSS"}: Ranked set sampling
}}

\item{confidence}{The confidence level to use.}

\item{replace}{Logical (default \code{TRUE}). Sample with replacement?}

\item{model_based}{An inference mode:
\itemize{
\item \code{FALSE}: design based inference
\item \code{TRUE}: model based inference using super population model
}}

\item{pop_size}{The population size. Must be provided if sampling without replacement, or if \code{model} is \code{1}.}
}
\value{
A \code{data.frame} with the point estimates provided by different types of estimators along with standard
error and confidence intervals.
}
\description{
Estimate means from RSS or JPS sample.
}
\examples{

# Compute the JPS estimators

jps_estimates <- rss_jps_estimate(
  data = emergence_ranks, set_size = 4,
  method = "JPS", confidence = 0.95,
  replace = TRUE, model_based = FALSE,
  pop_size = nrow(population)
)

}
