% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_sample.R
\name{rss_sample_wo_replacement}
\alias{rss_sample_wo_replacement}
\title{Generate ranked set sampling (RSS) without replacement on the population provided.}
\usage{
rss_sample_wo_replacement(pop, n, H, K)
}
\arguments{
\item{pop}{Population that will be sampled with an auxiliary parameter in the second column.}

\item{n}{Sample size.}

\item{H}{Set size for each ranking group.}

\item{K}{Number of rankers.}
}
\value{
A matrix with ranks from each ranker.
}
\description{
Generate ranked set sampling (RSS) without replacement on the population provided.
}
\keyword{internal}
