% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSDF.R
\name{RSSDF}
\alias{RSSDF}
\title{Generate ranked set sampling (RSS) with replacement on the population provided.}
\usage{
RSSDF(pop, n, H, K)
}
\arguments{
\item{pop}{Population that will be sampled with an auxiliary parameter in the second column.}

\item{n}{Sample size.}

\item{H}{Set size for each raking group.}

\item{K}{Number of rankers.}
}
\value{
A matrix with ranks from each ranker.
}
\description{
Generate ranked set sampling (RSS) with replacement on the population provided.
}
