% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSED2F.R
\name{RSSED2F}
\alias{RSSED2F}
\title{This function Computes RSS estimator and its variance without replacement sampling}
\usage{
RSSED2F(RV, Y, H, N)
}
\arguments{
\item{RV}{Rank values}

\item{Y}{Response variable}

\item{H}{Set size}

\item{N}{Population size}
}
\description{
This function Computes RSS estimator and its variance without replacement sampling
}
\keyword{internal}
