% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSEF.R
\name{RSSEF}
\alias{RSSEF}
\title{This function provides estimator for RSS data}
\usage{
RSSEF(data, set_size, replace, model, N, alpha)
}
\arguments{
\item{data}{An n by (K+1) dimensional data matrix, the first column is Y-values, the next K coulumns are the ranks of K-ranking methods}

\item{set_size}{The set size}

\item{replace}{Logical. Sample with replacement?}

\item{model}{If Modle=0, use design based inference, if model=1, use superpopulation model}

\item{N}{The population size}

\item{alpha}{The significance level}
}
\description{
This function provides estimator for RSS data
}
\keyword{internal}
