% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbs_pps_heatmap.R
\name{sbs_pps_heatmap}
\alias{sbs_pps_heatmap}
\title{Generate a heat map of SBS PPS sample of the provided population.}
\usage{
sbs_pps_heatmap(pop, sbs_indices, pps_indices)
}
\arguments{
\item{pop}{Population data frame to be sampled with 5 columns.
\enumerate{
\item Halton numbers
\item X1-coordinate of population unit
\item X2-coordinate of population unit
\item Size measurements of population units
\item Inclusion probabilities
}}

\item{sbs_indices}{Indices of SBS sample.}

\item{pps_indices}{Indices of PPS sample.}
}
\value{
Heat map of the sample.
}
\description{
Generate a heat map of SBS PPS sample of the provided population.
}
