% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbs_pps_sample.R
\name{sbs_pps_sample}
\alias{sbs_pps_sample}
\title{Generate probability proportional to size (PPS) and spatially balanced sampling on the population provided.}
\usage{
sbs_pps_sample(population, n, parallelize = TRUE)
}
\arguments{
\item{population}{Population data frame to be sampled with 5 columns.
\enumerate{
\item Halton numbers
\item X1-coordinate of population unit
\item X2-coordinate of population unit
\item Size measurements of population units
}}

\item{n}{Sample sizes (SBS sample size, PPS sample size).}

\item{parallelize}{A flag whether to parallelize the computational task.}
}
\value{
A named list of:
\itemize{
\item heatmap: heat map of the sample
\item sample: SBS PPS sample of the population
}
}
\description{
Generate probability proportional to size (PPS) and spatially balanced sampling on the population provided.
}
