% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_sample.R
\name{single_bootstrap}
\alias{single_bootstrap}
\title{Generate a single boostrap sample on the provided population.}
\usage{
single_bootstrap(pop, n)
}
\arguments{
\item{pop}{Population data}

\item{n}{Sample sizes (SBS sample size, PPS sample size).}
}
\value{
A summary data frame of the estimator.
}
\description{
Generate a single boostrap sample on the provided population.
}
