% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRIPLEF.R
\name{TRIPLEF}
\alias{TRIPLEF}
\title{Calculate the triple sum}
\usage{
TRIPLEF(uv, H, n)
}
\arguments{
\item{H}{The set size}

\item{n}{The sample size}
}
\description{
Calculate the triple sum
}
\keyword{internal}
