% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WEIGHTF.R
\name{WEIGHTF}
\alias{WEIGHTF}
\title{Calculate Agreement Weights}
\usage{
WEIGHTF(RV, set_size)
}
\arguments{
\item{RV}{Ranking values}

\item{set_size}{Set size}
}
\value{
A vector of agreement weights
}
\description{
Calculate Agreement Weights
}
\keyword{internal}
