% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JPSD2F.R
\name{JPSD2F}
\alias{JPSD2F}
\title{Title}
\usage{
JPSD2F(pop, n, H, tau, K)
}
\arguments{
\item{pop}{Population}

\item{n}{sample size}

\item{H}{Set size}

\item{tau}{controls the ranking quality}

\item{K}{Number of rankers}
}
\value{

}
\description{
Title
}
