% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JPSED0F.R
\name{JPSED0F}
\alias{JPSED0F}
\title{This function Computes JPS estimator and its variance}
\usage{
JPSED0F(RV, Y, set_size, Coef, N, Replace, Model)
}
\arguments{
\item{RV}{Ranking values provided}

\item{Y}{The data values that were ranked}

\item{set_size}{The set size}

\item{Coef}{Coefficients}

\item{N}{Population size}

\item{Replace}{Logical. Sample with replacement?}

\item{Model}{If model is 0, it's design based inference, if model = 1, it is model based inference using super population model.}
}
\value{
A vector with two elements: the mean and variance estimate from JPS
}
\description{
This function Computes JPS estimator and its variance
}
\keyword{internal}
