% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JPSEDF.R
\name{JPSEDF}
\alias{JPSEDF}
\title{Computes the estimator and variance for each individual ranker}
\usage{
JPSEDF(RV, Y, H, N, Coef, CoefDel, Replace, Model, K)
}
\arguments{
\item{RV}{Rank values for Y}

\item{Y}{Response measurements}

\item{K}{}
}
\value{

}
\description{
Computes the estimator and variance for each individual ranker
}
\keyword{internal}
