% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JPSEDF.R
\name{JPSEDF}
\alias{JPSEDF}
\title{Title}
\usage{
JPSEDF(RV, Y, H, N, Coef, CoefDel, Replace, Model, K)
}
\arguments{
\item{RV}{Rank values for Y}

\item{K}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
