% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JPSEF.R
\name{JPSEF}
\alias{JPSEF}
\title{Computes the estimator for JPS data}
\usage{
JPSEF(data, set_size, replace = TRUE, model, N, alpha)
}
\arguments{
\item{data}{The data to use for estimation.}

\item{set_size}{The set size.}

\item{replace}{Logical (default \code{TRUE}). Sample with replacement?}

\item{model}{If model is 0, it's design based inference, if model = 1, it is model based inference using super population model.}

\item{N}{The population size.}

\item{alpha}{The significance level.}
}
\value{

}
\description{
Computes the estimator for JPS data
}
\keyword{internal}
