% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JPSLF.R
\name{JPSLF}
\alias{JPSLF}
\title{Title}
\usage{
JPSLF(data, set_size, replace, population_size, Model)
}
\arguments{
\item{replace}{Are}

\item{population_size}{Population size}

\item{Model}{}
}
\value{

}
\description{
Title
}
