% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_Sample.R
\name{OneSample}
\alias{OneSample}
\title{Title}
\usage{
OneSample(
  data,
  set_size,
  method = "JPS",
  confidence = 0.95,
  replace = TRUE,
  model = 0,
  N = NULL
)
}
\arguments{
\item{method}{Takes values \code{JPS} (the default) for Judgement Post Stratification, or \code{RSS} for Ranked Set Sample.}

\item{confidence}{The confidence level to use. Defaults to 0.95}

\item{replace}{Logical (default \code{TRUE}). Sample with replacement?}

\item{N}{}
}
\value{

}
\description{
Title
}
