% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_Sample.R
\name{OneSample}
\alias{OneSample}
\title{Title}
\usage{
OneSample(
  data,
  set_size,
  method = c("JPS", "RSS"),
  confidence = 0.95,
  replace = TRUE,
  model = 0,
  pop_size = NULL
)
}
\arguments{
\item{data}{The data for \code{JPS} or \code{RSS} sampling}

\item{set_size}{The set size of the ranks.}

\item{method}{Takes values \code{JPS} (the default) for Judgment Post Stratification, or \code{RSS} for Ranked Set Sample.}

\item{confidence}{The confidence level to use.}

\item{replace}{Logical (default \code{TRUE}). Sample with replacement?}

\item{model}{If model is 0, it's design based inference, if model = 1, it is model based inference using super population model}

\item{pop_size}{The population size. Must be provided if sampling without replacement, or if \code{model} is set to 'superpopulation'.}
}
\value{
A \code{data.frame} with the rankings
}
\description{
Title
}
