% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_estimate.R
\name{rss_estimate}
\alias{rss_estimate}
\title{This function provides an estimator for RSS data}
\usage{
rss_estimate(data, set_size, replace, model_based, pop_size, alpha)
}
\arguments{
\item{data}{A matrix with a response variable in the first column and ranks in the following columns.}

\item{set_size}{Set size for each ranking group.}

\item{replace}{A boolean which specifies whether to sample with replacement or not.}

\item{model_based}{An inference mode:
\itemize{
\item \code{FALSE}: design based inference
\item \code{TRUE}: model based inference using super population model
}}

\item{pop_size}{The population size. Must be provided if sampling without replacement, or if \code{model} is \code{1}.}

\item{alpha}{A significance level.}
}
\description{
This function provides an estimator for RSS data
}
\keyword{internal}
