% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_sample.R
\name{rss_sample}
\alias{rss_sample}
\title{Generate ranked set sampling (RSS) on the population provided.}
\usage{
rss_sample(pop, n, H, K, with_replacement = FALSE)
}
\arguments{
\item{pop}{Population that will be sampled with an auxiliary parameter in the second column.}

\item{n}{Sample size.}

\item{H}{Set size for each raking group.}

\item{K}{Number of rankers.}

\item{with_replacement}{A boolean which specifies whether to sample with replacement or not.}
}
\value{
A matrix with ranks from each ranker.
}
\description{
Generate ranked set sampling (RSS) on the population provided.
}
