% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbs_pps_estimate.R
\name{sbs_pps_estimate}
\alias{sbs_pps_estimate}
\title{Compute an estimator for SBS PPS sampled data.}
\usage{
sbs_pps_estimate(
  population,
  n,
  y,
  sample_matrix,
  n_bootstraps = 100,
  alpha = 0.05,
  n_cores = getOption("n_cores", 1)
)
}
\arguments{
\item{population}{Population data frame to be sampled with 4 columns.
\enumerate{
\item Halton numbers
\item X1-coordinate of population unit
\item X2-coordinate of population unit
\item Size measurement of population unit
}}

\item{n}{Sample sizes (SBS sample size, PPS sample size).}

\item{y}{Sample response values.}

\item{sample_matrix}{Sample data frame to be sampled with 6 columns.
\enumerate{
\item Halton numbers
\item X1-coordinate of population unit
\item X2-coordinate of population unit
\item Size measurement of population unit
\item Weight
\item Inclusion probability
}}

\item{n_bootstraps}{Number of bootstrap samples.}

\item{alpha}{The significance level.}

\item{n_cores}{The number of cores to be used for computational tasks (specify 0 for max).}
}
\value{
A summary data frame of the estimator.
}
\description{
Compute an estimator for SBS PPS sampled data.
}
