% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_stage_cluster_sample.R
\name{two_stage_cluster_sample}
\alias{two_stage_cluster_sample}
\title{Generate two-stage cluster sampling on the population provided.}
\usage{
two_stage_cluster_sample(pop, sampling_strategies, n, H, ni, Hi)
}
\arguments{
\item{pop}{Population that will be sampled with these ordered columns:
\enumerate{
\item Parent id: an index to denotes the parent of the record
\item Parent auxiliary parameter: an auxiliary parameter for ranking parents
\item Child auxiliary parameter: an auxiliary parameter for ranking children
}}

\item{sampling_strategies}{(first stage sampling strategy, second stage sampling strategy), e.g.,
\code{c('srs', 'jps')}.
\itemize{
\item \code{'srs'}: simple random sampling without replacement
\item \code{'jps'}: JPS sampling without replacement
\item \code{'jps_replace'}: JPS sampling with replacement
}}

\item{n}{Number of samples to be ranked in the first stage.}

\item{H}{Set size for each ranking group in the first stage.}
}
\value{
A matrix with ranks from each ranker.
}
\description{
Generate two-stage cluster sampling on the population provided.
}
\examples{
set.seed(112)
population_size <- 600
# the number of samples to be ranked in each set
H <- 3

replace <- FALSE
sigma <- 4
mu <- 10
n_rankers <- 3
# sample size
n <- 9

population <- qnorm((1:population_size) / (population_size + 1), mu, sigma)
rho <- 0.75
tau <- sigma * sqrt(1 / rho^2 - 1)
x <- population + tau * rnorm(population_size, 0, 1)

population <- cbind(population, x)
rss_sample(population, n, H, n_rankers, replace)
#>            [,1] [,2] [,3] [,4]
#>  [1,]  8.910625    1    3    1
#>  [2,] 12.317145    2    2    1
#>  [3,] 11.619746    3    3    2
#>  [4,]  8.307549    1    2    1
#>  [5,]  5.089992    2    2    3
#>  [6,]  7.233575    3    3    3
#>  [7,] 11.601654    1    2    2
#>  [8,]  9.134107    2    1    1
#>  [9,] 12.960431    3    3    1

}
